FingerTree

Description

Library implementing FingerTree, a general purpose data-structure, which can be easily specialized to implement other data structures.

amortized O(1) access to and manipulation with first and last element

O(log n) access to other elements

O(log n) concatenation

O(n) construction from list and deconstruction to list

FingerTree

data FingerTree m a

General purpose tree-like structure

Constructors

Empty
no elements

Single a
a single element

Deep m (Vec m a) (FingerTree m (Vec m a)) (Vec m a)
tree-node with nonempty b+ tree on each side and pointer to next layer, whose b+ trees are one level deeper

data Vec m a

list with length tracking and annotation, which provides auxiliary information about its elements

Constructors

Vec

Fields

annotation :: m

len :: Int

elems :: [a]

Operations

class Structure (s :: * -> * -> *) where

Minimal complete definition

(<|), (|>), popRSafe, popLSafe, mapS, traverseS

Methods

(<|) :: Measured a m => a -> s m a -> s m a infixr 5

Adds element at the beginning of the Structure

(|>) :: Measured a m => s m a -> a -> s m a infixl 5

Adds element at the end of the Structure

popRSafe :: Measured a m => s m a -> Maybe (a, s m a)

Safely remove element from the end of the structure and return it along with new Structure

popLSafe :: Measured a m => s m a -> Maybe (a, s m a)

Safely remove element from the beginning of the structure and return it along with new Structure

mapS :: (Measured a m, Measured b n) => (a -> b) -> s m a -> s n b

Maps structure while changing annotation

Note: Haskell unfortunately doesn't permit anything like 
 (Measured a m, Measured b m) => Functor (FingerTree m)  
 to make FingerTree a Functor whenever both a and b can be measured, 
 so general Functor instace needs to be instantiated separately for each concrete implemenatation

traverseS :: (Measured a m, Measured b n, Applicative f) => (a -> f b) -> s m a -> f (s n b)

Traverses structure while changing annotation

popL :: Measured a m => s m a -> (a, s m a)

Removes element from beginning of Structure and returns it along with new Structure

popR :: Measured a m => s m a -> (a, s m a)

Removes element from end of Structure and returns it along with new structure. Throws error if unable to retrieve any element

getLSafe :: Measured a m => s m a -> Maybe a

Safely get the first element of the Structure

getRSafe :: Measured a m => s m a -> Maybe a

Safely get the first element of the Structure

getL :: Measured a m => s m a -> a

Get the first element of the Structure.Throws error if unable to retrieve any element

getR :: Measured a m => s m a -> a

Get the last element of the Structure.Throws error if unable to retrieve any element

bind :: Measured a m => FingerTree m a -> (a -> FingerTree m a) -> FingerTree m a

Monadic bind that requires measured instance

O(n)

concatTrees

Arguments

:: Measured a m

=> FingerTree m a
tree1

-> [a]
elements to put between the two trees

-> FingerTree m a
tree 2

-> FingerTree m a

Concatenates two trees together and puts additional elements in the middle

O(n)

best case O(log n) when number of elements in middle is asymptoticaly irelevant

update

Arguments

:: Measured a m

=> (m -> Bool)
annotation predicate

-> (a -> a)
updating function

-> FingerTree m a
structure to update

-> (a, FingerTree m a)
(oldValue, newFingerTree)

Update first element which satisfies annotation predicate

O(log n)

replace

Arguments

:: Measured a m

=> (m -> Bool)
predicate

-> a
new value

-> FingerTree m a
structure

-> (a, FingerTree m a)
(oldValue, newFingerTree)

Replaces first element satifying annotation predicate

O(log n)

deep :: Measured a m => Vec m a -> FingerTree m (Vec m a) -> Vec m a -> FingerTree m a

Smart constructor for Deep which automaticaly resolves annotation

O(1)

Applications

Order based FingerTrees

type Heap a = FingerTree (Minimax a) (Value a)

type PriorityQueue a = FingerTree (Minimax Int) (Prioritized a)

type IntervalTree = FingerTree (Minimax Int) Interval

popMax :: (Ord a, Ord b, Measured a (Minimax b)) => FingerTree (Minimax b) a -> (a, FingerTree (Minimax b) a)

Removes maximal element and returns it along with modified tree

O(log n)

popMin :: (Ord a, Ord b, Measured a (Minimax b)) => FingerTree (Minimax b) a -> (a, FingerTree (Minimax b) a)

Removes minimal element and returns it along with modified tree

O(log n)

overlapInclusive :: Interval -> IntervalTree -> Interval

Finds first occurence in tree which ovelaps with supplied inclusive interval

O(log n)

overlapExclusive :: Interval -> IntervalTree -> Interval

Finds first occurence in tree which ovelaps with supplied exclusive interval

O(log n)

Indexed FingerTrees

type Seq a = FingerTree Size (Value a)

type SegmentTree a = FingerTree (Segment a) (Value a)

class Indexed i where

Minimal complete definition

getSize

Methods

getSize :: i -> Int

Locates size inside a datatype

adjust :: Measured a i => (a -> a) -> Int -> FingerTree i a -> FingerTree i a

Adjusts n-th element

(!) :: Measured a i => FingerTree i a -> Int -> a

Gets n-th element

replaceAt :: Measured a i => a -> Int -> FingerTree i a -> (a, FingerTree i a)

Replaces n-th element

Other

type SimpleFingerTree a = FingerTree () (Value a)

class (Monoid m, Eq m) => Measured a m where

Methods

measure :: a -> m

data Size

data Value a

data Interval

Interval with start and end

data Segment a

Tracking of size and sum

data Max a

tracking of minimum with monoidical identity

data Min a

tracking of maximum with monoidical identity

data Minimax a

Tracking of both maximum and minimum

data Prioritized a

Data with priority